# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  title: string;
  category?: string | null;
  imagePath: string;
  createdAt: Date;
}